<fix>: move post-like entity to proper directory
