+++ src/app.module.ts
@@ -9,6 +9,7 @@ import { ProjectModule } from './project/project.module';
 import { UserProjectModule } from './user-project/user-project.module';
 import { DeserializeUserMiddleWare } from './common/middleware/deserialize-user.middleware';
 import { PostModule } from './post/post.module';
+import { PostLikeModule } from './post-like/post-like.module';
 
 @Module({
   imports: [
@@ -18,6 +19,7 @@ import { PostModule } from './post/post.module';
     ProjectModule,
     UserProjectModule,
     PostModule,
+    PostLikeModule,
   ],
   controllers: [AppController],
   providers: [AppService],
